firstasdasd27

+++ server.js
@@ -85,7 +85,7 @@ function request_page(url, width_1, height_1, callback){
 					
 					var height_x = getQueryVariable('height',page.url);
 					
-					page.clipRect = { top: 0, left: 0, width: 1400, height: properties.height1, };
+					page.clipRect = { top: 0, left: 0, width: 1400, height: (int)properties.height1, };
 					
 					//page.clipRect = { top: 0, left: 0, width: 1400, height: 666, };
